First and main module done
AppPermissions, set a default policy (deny or allow), and exceptions to it. Also, module infra structure was increased, with abstracted methods in lib.modules to handle the modules. Increase on interface, and bugfixes. Implemented update dialog in Main window (if you open a file outdated (modules of file dont corresponding with modules in system)).